Restructure models by splitting each into base and constructors.
