+++ lib/auth.js
@@ -1,17 +1,18 @@
 import passport from 'passport'
 import { Strategy as LocalStrategy } from 'passport-local'
 
-import User from '../model/user/user'
-let userModel = new User()
+import UserBase from '../model/user/user.base'
+import { checkPassword } from '../model/user/user.helper'
+let userBase = new UserBase()
 
 passport.use(new LocalStrategy({usernameField: 'email'}, (username, password, done) => {
-  let user_data = userModel.get(username)
+  let user_data = userBase.get(username)
   if (!user_data) {
     return done(null, false, {
       message: 'User not found'
     })
   }
-  if (userModel.checkPassword(user_data.id, password)) {
+  if (checkPassword(user_data.id, password)) {
     return done(null, false, {
       message: 'Incorrect password'
     })
@@ -24,7 +25,7 @@ passport.serializeUser((user, done) => {
 })
 
 passport.deserializeUser((id, done) => {
-  userModel.get(id).then((user, err) => {
+  userBase.get(id).then((user, err) => {
     return done(err, user)
   })
 })
